maint: port to gcc -std=c99 -Wstrict-prototypes
This doesn’t fix any bugs, since the code is correct as-is on all
C versions.  However, it pacifies gcc -std=c99 -Wstrict-prototypes,
which complains about a function definition without arguments
unless there’s a previous prototype with ‘(void)’.
Most (but not all) of this patch merely reverts some recent
changes to be more consistent when porting to C23.
* dfltcc.c (is_dfltcc_enabled):
* gzip.c (get_input_size_and_time):
* inflate.c (inflate_stored, inflate_fixed, inflate_dynamic):
* unpack.c (read_byte):